Restructure app architecture
